todo_drf_react/api/serializers.py
from re import T
from rest_framework import serializers
from .models import Task

class TaskSerializers(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ('id','title','completed',)
